Make Finder and MatchedPath private
These data don't have to be exposed as public implementations because I
don't plan to provide this crate as a library.
