+++ src/finder.rs
@@ -4,14 +4,14 @@ use std::path::{Path, PathBuf};
 use crate::error::{Error, Result};
 use crate::matched_path::MatchedPath;
 
-pub struct Finder {
+pub(crate) struct Finder {
     starting_point: String,
     query: String,
     dirs: VecDeque<ConsumedDir>,
 }
 
 impl Finder {
-    pub fn new(starting_point: &str, query: &str) -> Result<Self> {
+    pub(crate) fn new(starting_point: &str, query: &str) -> Result<Self> {
         let mut dirs = VecDeque::with_capacity(100);
         let consumed_dir = ConsumedDir::new(starting_point)?;
         let starting_point = consumed_dir.root.clone();
@@ -109,4 +109,106 @@ fn canonicalize_starting_point(path: &Path) -> Result<PathBuf> {
             path,
         ))
     )
+}
+
+#[cfg(test)]
+mod tests {
+    use std::fs::{create_dir_all, File};
+    use std::io;
+
+    use tempfile::{tempdir, TempDir};
+
+    use super::*;
+
+    pub fn create_tree() -> io::Result<TempDir> {
+        let tmp = tempdir()?;
+        create_dir_all(tmp.path().join("src/a/b/c"))?;
+        create_dir_all(tmp.path().join("lib/a/b/c"))?;
+        create_dir_all(tmp.path().join(".config"))?;
+        let _ = File::create(tmp.path().join(".browserslistrc"))?;
+        let _ = File::create(tmp.path().join(".config/bar.toml"))?;
+        let _ = File::create(tmp.path().join(".config/ok.toml"))?;
+        let _ = File::create(tmp.path().join(".editorconfig"))?;
+        let _ = File::create(tmp.path().join(".env"))?;
+        let _ = File::create(tmp.path().join(".env.local"))?;
+        let _ = File::create(tmp.path().join(".gitignore"))?;
+        let _ = File::create(tmp.path().join(".npmrc"))?;
+        let _ = File::create(tmp.path().join(".nvmrc"))?;
+        let _ = File::create(tmp.path().join("Dockerfile"))?;
+        let _ = File::create(tmp.path().join("LICENSE"))?;
+        let _ = File::create(tmp.path().join("README.md"))?;
+        let _ = File::create(tmp.path().join("lib/a/b/c/index.js"))?;
+        let _ = File::create(tmp.path().join("lib/a/b/c/☕.js"))?;
+        let _ = File::create(tmp.path().join("lib/a/b/index.js"))?;
+        let _ = File::create(tmp.path().join("lib/a/index.js"))?;
+        let _ = File::create(tmp.path().join("lib/bar.js"))?;
+        let _ = File::create(tmp.path().join("lib/index.js"))?;
+        let _ = File::create(tmp.path().join("package-lock.json"))?;
+        let _ = File::create(tmp.path().join("package.json"))?;
+        let _ = File::create(tmp.path().join("src/a/__test__.js"))?;
+        let _ = File::create(tmp.path().join("src/a/b/c/index.js"))?;
+        let _ = File::create(tmp.path().join("src/a/b/index.js"))?;
+        let _ = File::create(tmp.path().join("src/a/index.js"))?;
+        let _ = File::create(tmp.path().join("src/a/☕.js"))?;
+        let _ = File::create(tmp.path().join("src/foo.js"))?;
+        let _ = File::create(tmp.path().join("src/index.js"))?;
+        let _ = File::create(tmp.path().join("tsconfig.json"))?;
+        let _ = File::create(tmp.path().join("☕.txt"))?;
+        Ok(tmp)
+    }
+
+    fn find_paths(starting_point: &str, query: &str) -> Vec<String> {
+        let mut paths = vec![];
+        for matched in Finder::new(starting_point, query).unwrap() {
+            let path = matched.unwrap();
+            paths.push(path);
+        }
+        let mut paths: Vec<String> = paths
+            .iter()
+            .map(|m| m.relative().replace('\\', "/"))
+            .collect();
+        paths.sort();
+        paths
+    }
+
+    #[test]
+    fn returns_all_paths() {
+        let dir = create_tree().unwrap();
+        let size = find_paths(dir.path().to_str().unwrap(), "").len();
+        assert_eq!(size, 29);
+    }
+
+    #[test]
+    fn returns_empty() {
+        let dir = create_tree().unwrap();
+        let size = find_paths(
+            dir.path().to_str().unwrap(),
+            "the word should be not found with 🎂",
+        )
+            .len();
+        assert_eq!(size, 0);
+    }
+
+    #[test]
+    fn returns_filtered_paths_with_only_separator() {
+        let dir = create_tree().unwrap();
+        let size = find_paths(dir.path().to_str().unwrap(), "/").len();
+        assert_eq!(size, 15);
+    }
+
+    #[test]
+    fn returns_filtered_paths_with_uppercase() {
+        let dir = create_tree().unwrap();
+        let paths = find_paths(dir.path().to_str().unwrap(), "licenSE");
+        assert_eq!(paths.len(), 1);
+        assert_eq!(paths, vec!["LICENSE"]);
+    }
+
+    #[test]
+    fn returns_filtered_paths_with_emoji_coffee() {
+        let dir = create_tree().unwrap();
+        let paths = find_paths(dir.path().to_str().unwrap(), "☕");
+        assert_eq!(paths.len(), 3);
+        assert_eq!(paths, vec!["lib/a/b/c/☕.js", "src/a/☕.js", "☕.txt"]);
+    }
 }
