+++ src/lib.rs
@@ -1,8 +1,5 @@
-pub use cli::entrypoint;
-pub use cli::safe_exit;
+pub use cli::{entrypoint, safe_exit};
 pub use error::{Error, ErrorKind, Result};
-pub use finder::Finder;
-pub use matched_path::MatchedPath;
 
 mod args;
 mod cli;
